refactor(cmd): 重构代码结构，将功能模块拆分到独立文件
将原cobra_cmd.go中的功能模块拆分为多个独立文件，包括messages.go、version_cobra.go、runssh_cmd.go和sync_cobra.go
优化代码组织结构，提高可维护性
